mfs: Avoid Operands don't affect result (coverity)
Should fix 362270

+++ src/dosext/mfs/mfs.c
@@ -477,7 +477,7 @@ done:
  *   redirector int2f/11 function. It relies on being able to run int2f/12
  *   functions which need to use the DOS stack.
  *****************************/
-static int GetCDSInDOS(uint8_t dosdrive, cds_t *cds)
+static int GetCDSInDOS(uint16_t dosdrive, cds_t *cds)
 {
   unsigned int ssp, sp;
   int ret = 1;
@@ -486,7 +486,7 @@ static int GetCDSInDOS(uint8_t dosdrive, cds_t *cds)
   /* Ask DOS for the CDS */
   ssp = SEGOFF2LINEAR(_SS, 0);
   sp = _SP;
-  pushw(ssp, sp, (uint16_t)dosdrive);
+  pushw(ssp, sp, dosdrive);
   _SP -= 2;
   _AX = 0x1217;
   do_int_call_back(0x2f);
@@ -2317,7 +2317,7 @@ debug_dump_sft(char handle)
 static int GetRedirection(struct vm86_regs *state, int rSize, int subfunc)
 {
   u_short index = WORD(state->ebx);
-  int dd;
+  unsigned int dd;
   u_short returnBX;		/* see notes below */
   u_short returnCX;
   u_short returnDX;
@@ -2456,7 +2456,7 @@ static int path_list_contains(const char *clist, const char *path)
  *   This function is used internally by DOSEMU, in contrast to
  *   RedirectDevice(), which must be called from DOS.
  *****************************/
-static int RedirectDisk(struct vm86_regs *state, int drive,
+static int RedirectDisk(struct vm86_regs *state, unsigned int drive,
     const char *resourceName)
 {
   char path[PATH_MAX];
@@ -2542,7 +2542,7 @@ static int RedirectDisk(struct vm86_regs *state, int drive,
 
 static int EnableDiskRedirections(void)
 {
-  int dd;
+  unsigned int dd;
   cds_t cds;
   int ret = FALSE;
 
@@ -2559,7 +2559,7 @@ static int EnableDiskRedirections(void)
 
 static int DisableDiskRedirections(void)
 {
-  int dd;
+  unsigned int dd;
   cds_t cds;
   int ret = FALSE;
 
@@ -2575,7 +2575,7 @@ static int DisableDiskRedirections(void)
 
 static int GetRedirModeDisk(void)
 {
-  int dd;
+  unsigned int dd;
   cds_t cds;
 
   for (dd = 0; dd < num_drives; dd++) {
